uid-eatcu/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
flask_login==0.4.1
werkzeug==0.15.2
flask_wtf==0.14.2
wtforms==2.2.1
wtforms-components==0.10.4
Flask-Mail==0.9.1
pytz==2019.1
